@W-22642116 Extract showError into its own module
Move showError, ERROR_UI, ERROR_HEADING, the Scenario type, and the SVG
import out of handleToolResult.ts into showError.ts; split its unit tests
into showError.test.ts. Pure refactor, no behavior change.
